+++ core/src/main/kotlin/de/otto/babbage/core/status/StatusData.kt
@@ -2,7 +2,7 @@ package de.otto.babbage.core.status
 
 import de.otto.babbage.core.status.contributors.SystemInfo
 import de.otto.babbage.core.status.indicators.StatusDetail
-import org.springframework.boot.actuate.health.Status
+import de.otto.babbage.core.status.version.GitVersion
 
 data class StatusData(
     val application: Application,
@@ -13,21 +13,5 @@ data class StatusData(
 
 data class Application(
     val name: String,
-    val version: String,
-    val status: ApplicationStatus,
-    val commit: String
+    val version: GitVersion? = null
 )
-
-enum class ApplicationStatus {
-    OK, WARNING, ERROR;
-
-    companion object {
-        fun fromHealth(status: Status): ApplicationStatus {
-            return when (status) {
-                Status.DOWN, Status.OUT_OF_SERVICE, Status.UNKNOWN -> ERROR
-                Status.UP -> OK
-                else -> ERROR
-            }
-        }
-    }
-}