+++ core/src/main/kotlin/de/otto/babbage/core/status/version/GitVersion.kt
@@ -1,0 +1,10 @@
+package de.otto.babbage.core.status.version
+
+data class GitVersion(
+    val branch: String,
+    val commitId: String,
+    val time: Long,
+    val user: String,
+    val message: String,
+    val repositoryUrl: String
+)